Ensure fullscreen and immersive VR video is exited on pause (#3182)
Fixes #3154

Previously if the content process was killed or crashed while the
application was paused, the windows would still be in fullscreen
and/or immersive VR video mode while the recreated GeckoSession
would not. This partially fixes the issue by ensuring fullscreen
and VR immersive video is exited on pause so the windows do not
get into a bad state if the crash/kills while the application is
backgrounded.