Privacy Policy dialog in VR
Display Wolvic's Privacy Policy the first time that the user launches
the application. If the user accepts it, we store a preference and
con't show the dialog again. If the user rejects it, we close the
application because we can not continue (although we should
probably ask the user to confirm their decision beforehand).

The Privacy Policy text is stored in a handful of string values that
are displayed in TextViews with some custom styling.