Don't rebuild the Search Engine UI if not needed
The Search Engine selector is updating its UI unnecessarily whenever the user selects a new value.

The problem is that we rebuild the UI whenever the stored preference changes, but this is not needed most of the time because the list of available search engines remains the same.

With this PR, the Search Engine UI will only rebuild its UI when the selected radio button does not match the current stored preference.