HVR: Add a workaround for 6DoF controllers' poses (#115)
The HVR OpenXR SDK incorrectly returns the same poses for both aim and grip for 6DoF controllers. That was causing Wolvic to render both the controllers and the beams pretty inconsistently in both the 3D browser and WebXR environments.

This issue will be reported upstream to the SDK folks but in the meantime we'll add a workaround which basically generates the aim pose by applying some rotation to the value returned by the runtime (the runtime is basically always returning the grip pose).

In order to get the 2D browser experience also correct, we had to rotate the controller 3D models by 45º. This is something that we're already doing for the Oculus ones as well, so bonus points for consistency.