Enable MSAA in WebXR sessions
Wolvic allows users to set the MSAA level in the settings. It's possible
to completely disable it or to set it to either 2x or 4x. Wolvic was
properly setting those MSAA levels when creating the Gecko session.
However those settings were not properly applied to WebGL canvases
because Gecko does not enable antialiasing in XRWebGLLayers unless
the webgl.msaa-force setting is enabled.

The name of the Gecko setting is indeed a bit missleading because
it isn't about forcing really, it's about enabling actually.

After this change we'd be able to enjoy antialiased WebXR sessions.