Added a new flavor dimension "country"
The HVR builds require some specific parameters for the packages to
be distributed in mainland China. We're adding a new dimension called
"country" to deal with that. So far we just need China and World
variants. This should only affect HVR builds.

As an example use the 2D UI only for non-China HVR builds