Use the external storage for downloads (#105)
Always use the external storage for downloads. Specifically, the folder returned by context.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS). It turns out that this is already the existing behaviour, because we were saving the downloaded files to the same folder regardless of the value of the downloads storage setting (this PR also removes that option).

We don't need additional permissions when writing to the location returned by getExternalFilesDir(), so those permission requests have been removed.