Enable MSAA in GeckoView (#118)
Wolvic allows users to set the MSAA level in the settings. It's possible
to completely disable it or to set it to either 2x or 4x. Wolvic was
properly setting those MSAA levels when creating the Gecko session.
However those settings were not properly applied to WebGL canvases
because Gecko does not enable antialiasing in XRWebGLLayers unless
the webgl.msaa-force setting is enabled. The name of the Gecko setting 
is indeed a bit missleading because it isn't about forcing really, it's about 
enabling actually.

From early testing Oculus Quest2 does fine with MSAA 2x. However there
is a significant drop in the performance of Huawei Glass devices, so
we keep it disabled by default. The users can always change it in the
Wolvic settings, that should work fine from now on.