Handle the crash reporting service work in the UI thread
The crash reporting service requires access to the GeckoRuntime. The runtime is usually
already created by the time this service executes, but in some cases (like at startup)
it might be the case that it hasn't been created yet. Creating the GeckoRuntime must be
done in the main (UI) thread, so we must ensure that we call
EngineProvider.getOrCreateRuntime() in the UI thread.