Migrate Wolvic to Gecko 103.0.2
The migration involves updating several APIs that were removed/deprecated
or changed. In particular:

* Replaced EXTRA_CRASH_FATAL (deprecated) by EXTRA_CRASH_PROCESS_TYPE
* Removed rawId which was removed from MediaSource
* Disable GPU process as it breaks WebXR
* Migrate surfaceChanged to new Builder API
* Remove method that was deprecated
* Update signature of GeckoSurfaceTexture::lookup

From now on Wolvic requires building against this branch
https://github.com/Igalia/gecko-dev/tree/FIREFOX_103_0_2_RELEASE
which is based on gecko-view 103.0.20220808125904

Using the maven build is also supported but WebXR won't work
as it requires some downstream changes in our fork that are still
not integrated upstream.