Use desktop overrides to set DesktopMode for specific streaming sites
Now that UAOverride was refactored in UriOVerride, and that trySetUaMode()
in in place, we can automatically select desktop mode for the sites
listed in the new desktopModeOverrides.json file. We do it in
Session::onLocationChange because that allows us to "capture" all the
load attempts no matter they come from user input, or navigation requests
from other sites (like searching in a web search engine).

Note that we only do that if we are not in desktop mode already and also
only if we come from another site. This allows users to manually disable
desktop mode for the sites in the json file if they wish for whatever reason.