Refactor Session::setUaMode()
We're extracting the first part of that method into a new one called
trySetUaMode() in order to have a specific method that does all the
work related to setting the desktop mode but without issuing a
loadUri() call. Once this is in place we could add the code that
selectively enables desktop mode for a specific set of sites.