Fix saving state of SessionState
This is a regression from the multi-backend patch. In order to save the SessionState
object we need to call toJson() instead of toString(). The confugion came
from the fact that before the multi-backend patch we were directly calling
GeckoSessionState.toString() which actually generates a JSON. But now we need to use
the WSessionState::toJson() method to achieve the same result.

This was detected because the GeckoSession was issuing "String does not represent
valid session state." errors on startup.