Early register the external context for GeckoView
As mentioned in the docs
https://searchfox.org/mozilla-central/source/mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoVRManager.java#13
the external context (ExternalVR) must be initialized before using Gecko. Among other things
Gecko could fail detecting XR runtimes if we try to load some URL before setting
the external context, for example when quickly restoring a session at start.

Moved the registration to BrowserWorld creation so that we don't depend on
specific implementations of the native libs.