Wait for media full screen availability before entering video VR mode
This fixes an issue which happened whenever the YouTube fullscreen button was clicked
before the video was available. The outcome was a fullscreen black window with no video
and no audio.

The problem was the following. NavigationBarWidget was receiving the onFullScreen() event
from the content delegate (not from the media delegate). That method was calling
enterVRVideo() using a delay (perhaps an older hacky attempt to "fix" this issue). Then
enterVRVideo() tried to get the fullscreen video by calling getSession().getFullScreenVideo().
However the fact that we get onFullScreen() from the content delegate does not imply that
the media element is in fullscreen yet. Actually in many cases it was not. The consequence was
that a later call to WindowWidget::setVisible(false) was making the Session to be marked
as inactive because mIsInVRVideoMode was false as enableVRVideoMode() had not been called yet
(because there was no getSession().getFullScreenVideo()). A inactive session is basically a
non visible session with almost no associated resources, that's why audio and video stopped.

The solution is to change enterVRVideo() so that we wait for the onFullScreen event
*from the Media object* in case getSession().getFullScreenVideo() returns null. By waiting
for the Media:onFullScreen event we can ensure that there is a valid fullscreen video.

Note that this solution is not perfect because there are some corner cases that do not work
perfectly from the UI POV. For example, sometimes we get the Media::onFullScreen event
with zero width&height from the video. Zero means unknown for Gecko. In those cases we just
use the size of the window. This might cause the video to be played with a low resolution.
Another non ideal outcome is that sometimes we never get the Media::onFullScreen event from
Gecko so video is played in a "big screen" instead of in fully immersive environment.