Fix a crash in onSlowScript handler
The onSlowScript() handler in ContentDelegateImpl was not properly handling the case of
the ContentDelegate returning null for the onSlowScript() method. As the Gecko API
perfectly states the return value for onSlowScript() is nullable so we must handle
that case properly.

The code was assuming that it was never null so that's why it ended up doing
a null pointer dereference.