Update the perspective projection after changing the IPD
The WebXR sessions in Quest2 were showing significant distortions
when rotating the head in the z axis if and only if the lenses were
in positions 1 or 3, which correspond to the largest and smallest
values for interpupillary distance.

Changing the interpupillary distance at hardware level requires
recomputing the perspective projection. We don't have to do it
on every single frame but just as a response to the change.