HVR: improve android activity lifecycle handling
Wolvic used to exit whenever we got a onSurfaceDestroyed() event. That
does indeed happen when the activity is destroyed, but it does also
happen in situations in which we don't want to exit Wolvic but just
put it "in background". Examples of those are the system showing a
confirmation dialog (for example a permission request dialog) or when
the user moves out of the security zone.

From now on, whenever the onSurfaceDestroyed event is received, we
just invoke xrRequestExitSession() so the XR session transitions to
STOPPED and EXITING states. The render loop is not exited but throttled,
meaning that instead of render anything it just activelly waits for
new events.

Later on, if the session is resumed (for example when the user
re-enters the security zone, or by closing a system permission dialog)
we just have to invoke xrBeginSession() to get the OpenXR machinery
working.

The problem is that when the user clicks on the menu button of the R
controller (in order to exit the application) we basically get the same
sequence of activity calls (onPause, onSurfaceDestroyed, onStop). So we
should come up with some way of differentiate the cases of exiting the
application and stopping the render loop for a while.

The HVR SDK does a pretty bad job in those situations. Instead of letting
the activities handle their termination, it basically just abruptly
terminate them and go back to the system environment. That's why we
had to add an extra JNI method to inform the native-lib that the
application is going to be destroyed. The native-lib then will use
that information in the next render loop cycle to terminate the render
thread and free all the allocated graphic resources.