Let the user pick the speech recognition service
Add a new section in the Settings to select the desired speech recognition service.

The services available for a particular build are specified at compilation time and will be listed in `BuildConfig.SPEECH_SERVICES`. The one in the first position will be the default.

The class SpeechServices encapsulates the necessary information to identify and instantiate each service.

New strings:

- `settings_language_choose_language_voice_search_service_title`
- `settings_language_choose_language_voice_search_service_description`
- `language_service_options_reset`
- `language_options_voice_search_service_title`
- `voice_service_metkai`
- `voice_service_huawei_asr`