Notify about video availability changed
WindowWidget sets itself as the media session video availability delegate. This means
that when a Media session is activated, it will call
WindowWidget::onVideoAvailabilityChanged. So far so good. The thing is that WindowWidget
defines the WindowListener interface with several methods. One of them is precisely
onVideoAvailabilityChanged. The problem was that when WindowWidget received the
onVideoAvailabilityChanged notification from Media it was not forwarding it
to its listeners. We just had to add a loop over the listeners calling the method.