Wait for media full screen availability before entering video VR mode
This fixes an issue which happened whenever the YouTube fullscreen button was clicked
before the video was available. The outcome was a fullscreen black window with no video
and no audio.

The problem was the following. NavigationBarWidget was receiving the onFullScreen() event
from the content delegate (not from the media delegate). That method was calling
enterVRVideo() using a delay (perhaps an older hacky attempt to "fix" this issue). Then
enterVRVideo() tried to get the fullscreen video by calling getSession().getFullScreenVideo().
However the fact that we get onFullScreen() from the content delegate does not imply that
the media element is in fullscreen yet. Actually in many cases it was not. The consequence was
that a later call to WindowWidget::setVisible(false) was making the Session to be marked
as inactive because mIsInVRVideoMode was false as enableVRVideoMode() had not been called yet
(because there was no getSession().getFullScreenVideo()). A inactive session is basically a
non visible session with almost no associated resources, that's why audio and video stopped.

The solution is to change NavigationBarWidget:onFullScreen() so that it does not enter the
fullscreen immersive mode in case we don't have a valid fullscreen active media available.
In those cases we start listening for two possible notifications:
    1). if the video is not active yet the we wait for onVideoAvailabilityChanged which will
    be called once the video is active. This case is not optimal with Gecko because we won't
    be ever notified about the actual fullscreen size of the video, so we'll have to fallback
    to using the resolution of the fullscreen window. This causes that sometimes we display
    low-res videos even though the selected resolution in the UI is higher.
    2. if the video is active but not in fullscreen, then we wait for onMediaFullscreen. This is
    a new event handler added to WindowWidget. It's different from onFullscreen because the
    former is triggered by the media element while the latter is triggered by the document.

This means that most of the code from onFullScreen() was moved to a new method enterFullScreen()
which does the same but that it's only called whenever we have a valid active media session.
With this new approach we should *always* get into video immersive mode. Sometimes it might take
1'-2' depending on the time required to load the video, but we should end up in immersive mode.