Use multiple threads to compress textures
We use a software called EtcTool to generate compressed .ktx textures
from diverse bitmap files, for example from the cubemap that generates
the skybox. That tool has a parametter that allows you to specify the
number of threads to be used in the compression. By default it's 1.
Since it's a very expensive operation we should try to squeeze the
computer to use all the available CPUs.

For reference, in a 20 core machine, generating the texture for the
Meta Quest2 controller goes from 9.02s to 3.66s.