Fix a crash when the selected language is removed from the list of supported languages
In the event of removing a language from the list of supported languages (unlikely)
Wolvic might crash with IllegalStateException being thrown because the index
for the language stored in Settings does no longer exist. This situation normally
should only happen while developing and testing code but it's better to be safe.
In the event of having this issue we can just reset to the default language.