Restore CDATA blocks of translations made in Weblate
Some of our strings contain embedded HTML, and these are wrapped
by a CDATA block in the original, monolingual languale file.

The Weblate service we started using for translations doesn't pass
these CDATA blocks to the translations, so they are lost and the
embedded HTML gets broken.

We don't have an automated solution for this yet, so for now we need
to fix the CDATA blocks manually.