Request WRITE_EXTERNAL_STORAGE from the Settings
The storage permission is needed for some functionality, for example taking screenshots in HVR.

When this happens, the user can grant the storage permission in the device's settings.

Alternatively, the user can change the "storage" switch in the "Privacy & Security" section of Wolvic's settings. This will request a system dialog to confirm the change.

These two methods are supposed to be equivalent, but they are not: the first one will grant the WRITE_EXTERNAL_STORAGE permission, but the second one will only request READ_EXTERNAL_STORAGE.

This PR requests the WRITE_EXTERNAL_STORAGE permission from Wolvic's settings.