Prevent a crash due to incompatible logins databases
Users with old versions of Wolvic installed who are migrating to 1.0 are
experiencing a crash due to incompatible Logins database. This service is
provided by a Mozilla component which may have changed to use a different
database format.

This patch serves as a temporary fix to avoid the crash, by clearing saved
logins database if an incompatibility is found.