Add Firefox Accounts domain to Desktop mode overrides
This is a temporary workaround to fix login into Firefox Accounts, which
has been broken unexpectedly by a change in the server behavior: it is
apparently filtering by UA and the usual request we make to the OAuth
page keeps loading forever, stalling the login process.

This patch forces domains where host is `accounts.firefox.com` to always
use Desktop mode.