Revert "Use eye transform instead of translation in external_vr"
This reverts commit bb776dd67d1b4deff272d347ab1136f1c13a9747.

This was an attempt to fix some distortions we were seeing in some devices.
However it didn't make any difference in the end as the problem was in some
other place. In practice it was changing nothing so we could live with it
but the fact is that it's a barrier for new developers because it forces
devs to build our fork of Gecko to support WebXR sessions (as we were
bumping the version of the protocol used by GeckoView).