Delete logins database file on DB corruption
Whenever the logins database gets corrupted (for unknown reasons at this point)
every attempt to access it (even wiping it) will result in an Exception being
thrown. That exception if not handled could result in a crash.

While mozilla-components do not provide a better fix all we can do is to delete
the database file (yeah really ugly) hoping for the best.