Add support for haptic feedback in OpenXR backend
For now we assume all controllers using the OpenXR backend have at
least one haptic actuator; which is the case for Huawei VR and Oculus.

We leave the frequency unspecified since the JS API doesn't provide
it. The OpenXR spec states that the runtime will choose an
appropriate value in this case.