Dismiss file dialog prompts by default
We don't support file dialogs yet, so dismiss the prompt request whenever
made without even showing any UI. This fixes a crash whenever a file prompt
was triggered because we were returning a null WResult in a place where it
was not expected.

This is a regression from the multi-backend patch.