Use a separate OnSharedPreferenceChangeListener
I don't know exactly the reason, but PlatformActivity is not able to receive updates from the SharedPreferences when the user accepts the legal terms on the phone UI.

However, a separate listener object inside PlatformActivity is able to receive those updates. PrivacyPolicyFragment is also able to receive the same updates.