Initial PushKit support (#213)
This implementation signs up in the Push Message service after the user accepts the Privacy Policy. After this registration is completed, incoming push messages will be processed at PlatformActivity.handleHmsMessageBroadcast().

Messages can be sent in two ways.

The first one is to follow the process at Notification Message Processing for Apps Running in the Foreground and include "foreground_show": false in the notification message. This means that when our app is in the foreground, the incoming message will be processed by it and not sent to Android's Notification Center. I could not test this because it requires access to the REST API.

The second way is to send a data message with a JSON payload. The format that I have followed is the same one that the documentation uses. For example:

{
    "message": {
        "notification": {
            "title": "The title of the message",
            "body":  "This is the body of the message."
        },
        "android": {
            "notification": {
                "foreground_show": false,
                "click_action": {
                    "type": 2,
                    "url": "https://wolvic.com"
                }
            }
        }
    }
}
The action is executed when the user taps the notification. In this case, it would open wolvic.com in a new tab.

It is also possible to open the main page of our app or launch a different app, but the documentation is a bit confusing on this point. As I understand it, this

{
    "message": {
       
        "android": {
            "notification": {
                "click_action": {
                    "type": 1,
                    "action": "android.intent.action.MAIN"
                    "intent": "intent://com.example/..."
                }
            }
        }
    }
}
should be implemented like this (but I am not completely sure):

Intent intent = new Intent( "android.intent.action.MAIN" );
intent.setData( Uri.parse( "intent://com.example/..." ) );
startActivity(intent);

I have added a new panel in the Library window to display incoming push messages. The UI is still clunky: when a new message arrives, we display a notification with its title but the user has to open the Library to see its content.

Incoming messages are not stored permanently, but they probably should be.