Filter out incompatible controllers based on DoF
In general each XR system is either 3DoF or 6DoF so it's enough
with having a single OpenXR input profile. That's why we used to
filter out mappings based on the name of the XR system.

However in the case of the HVR platform, you can easily convert
your 3DoF system in a 6DoF one by attaching a 6DoF kit to the
glasses. This means that filtering by name is not enough. We should
also have a way to specify different mappings for the same system
depending on whether they're either 3DoF or 6DoF.

This allows us to remove an #ifdef in the mappings. This will allow
us to eventually remove one of the two HVR flavours.