Make BUTTON_APP exit fullscreen mode (#278)
BUTTON_APP (the hamburger button on the left controller on Meta Quest2 and HVR)
is used to exit WebXR sessions and go back to the 2D browser. We can export that
very same behaviour to the immersive 360 videos as they're in a way a similar
experience.

In the particular case of the OculusVR platform, that button was mapped to the
Y hardware button when not in immersive mode. We're removing that as that
behaviour was not shared by the OpenXR backend. Also the Meta browser does
not do anything similar (the Y button is used to select the word that is pointed
by the aim ray). So in order to unify the behaviour of the different backends
we're now also treating it as a back navigation/exit XR/exit fullscreen button.

This fixes issues #275 and #276.