Remove the mapping for /input/home in HVR 6DoF controllers
There is no point in having a mapping for that button (in our case
it was mapped to BUTTON_APP) because that button is captured by the
system. The default (and only) action is to close the currently run
application, and there is nothing that apps can do ATM to change that.

That was causing, that sometimes when the user clicked that button
a glimpse of a confirmation dialog was shown while the application was
exiting.