[HVR] Exit app when clicking exit on the confirmation dialog (#289)
For Pico and HVR builds we show an exit confirmation dialog whenever
back action is triggered and we cannot go backwards in history. We
don't do that in Meta for example because the system does it for us.

In the case of HVR clicking on exit was not doing anything because
the onBackPressed() method was not implemented for that platform. We
do invoke finish now and exit the application properly.