[HVR] Do not exit the OpenXR session when exiting the security zone
When we exit the security zone defined in HVR platform the system
destroys the android surface we use to render. The code in response
was terminating the OpenXR session, moving out of the render loop,
and exiting assuming that the application was being closed.

In order to properly handle it we don't exit the render loop
immediately now. We just throttle it, and wait until the user
re-enters the security zone. Whenever that happens the system
creates another surface for us, and we just need to call
xrBeginSession and resume the render loop.

The problem is that due to https://github.com/hms-ecosystem/OpenXR-SDK/issues/43
we cannot differentiate between the cases of exiting the security
zone and closing the application, since the HVR platform does
not properly implement the "exit to lobby" action when the menu
button of the right 6DoF controller is pressed (it just destroys
the surface instead of pausing, stopping and destroying the
activity). That's why we have to re-introduce the mapping
for the menu button of the right controller.

But now instead of doing a back operation, we basically call
finish(). By doing that the application is properly terminated
and resources are freed. The render loop is also properly
finalized in that case.

Last but not least, this fix was causing crashes in Android
rendering code in Harmony 2.0 but seems to work perfectly in
Harmony 3.0.