[HVR] Remove the hvr3dof flavor
After properly removing all the #ifdefs for the 3DoF specific
behaviours, we can finally remove the hvr3dof build flavor and
generate a single APK that supports both 3dof and 6dof controllers.

The AndroidManifest.xml file for the 3DoF flavour was merged with
the 6DoF one. In particular there were a couple of queries needed
in order to use the 3DoF controller instead of the phone control.