[HVR] Enable antialiasing by default for Harmony OS 3.0
Looks like the Harmony OS 3.0 gfx capabilities are much better than
in 2.0. We'll tempatively enable MSAA 2x by default also for the
HVR platform.

As we are deciding that at runtime now there is no point in having
a build flag for the default MSAA level and thus, we removed that.