[HVR] Do not exit app when going back in immersive mode
This is a regression of the patch that allowed Wolvic not to close
when exiting the security zone. The menu button of the R controller
was mapped to BUTTON_APP in order to properly detect it and exit
the render loop. The problem is that BUTTON_APP is also used to
simulate a back action. That's why clicking on the L menu button in
immersive experiences was making Wolvic to exit instead of just
closing the WebXR immersive session.

The solution is to keep track when the user presses that button
and exit the application if we get the onStop call (we could also
get the onStop call when exiting the security zone but in that case
we won't get the button press, so we won't exit).