Dim background when entering Video fullscreen mode with a fading effect
This was already implemented but not working fine, because there were
two calls dimming the background, the first one setting the proper
value and the second one to a very low dimming effect. By removing
the extra one we could have a very nice fade in/out dimming effect when
entering/exiting video fullscreen mode.

Apart from that the length of the animation was increased from 40 to
100 as it was way too fast.