Improve background dimming when entering video fullscreen mode
This was already implemented but it looked like not working because
the default selected alpha for dimming was 50% which is not that
easy to perceive. The fullscreen mode already implements 3 default
brightness levels: 100%, 50% and 0%, for normal, dark and void. We
have replaced them by 50%, 2.5% and 0%.

Apart from that we're removing an extra call to pushWorldBrightness
that happened before the pushWorldBrightness triggered by the
brightness widget as it makes no sense to do it twice. Same happened
with a popWorldBrightness done before the one done by the same
brightness widget.

Last but not least, the length of the animation was increased from
40 to 100 as it was too fast.