Upgraded AC to 75.0.22
So far, we have being using v63 which is rather old. For now we're
sticking to v75 because starting in v90, there are several
disruptive changes in the browser search component that require
an extra good amount of work. This should be seen as a good first
step instead of a final solution.

Note that the Mozilla appservices to be used with this specific
version is indeed 74.0.1. That is not a mistake, both versions
differ in the actual figure but are synchronized.

This allows us to remove the previous attempts to deal with
the login database corruption that was causing crashes on
startup (see
https://github.com/mozilla-mobile/android-components/issues/6681#issuecomment-1285773507)