Add new favicons as we browse the Web
Link the `BrowserIcons` instance with its associated extension, so the list of known favicons can be updated as we browser the Web.

This change required adding the `Engine` interface to `WolvicWebExtensionRuntime` so we can call `BrowserIcons.install()`.

`Engine` inherits from `WebExtensionRuntime`, which is what ` WolvicWebExtensionRuntime` implements. Unfortunately, the component's API requires an instance of the child interface, even if it does not actually use any functionality that is specific to it.