Parse and store Web apps
When the current website contains a link to a Web app manifest, we will receive this manifest (in JSON format) at Session.onWebAppManifest().

The manifest describing a Web app for the current website is encapsulated in a WebApp object and stored in SessionState.

The user can keep a list of installed Web apps. This list is saved in the shared preferences in JSON format. The WebAppsStore class can be used to access and modify this list.

An instance of WebAppsStore is kept in SessionStore.