Add support for showing/hiding a controller model
This wraps the controller model node by a toggle node; and adds a
SetModelVisible() method to controller delegate and controller container,
to show/hide the controller model.

Since we are reusing controller infrastructure for hand-tracking
support, this is useful to hide the controller models while still
showing the beam and pointer target.