OpenXR: Emulate Oculus Touch interaction profile if hand tracking is enabled
Since we are emulating controller functionality for hand tracking, we
need to always have an active interaction profile, even if Wolvic is
launched without any controllers active.

This patch emulates the activation of OculusTouch controllers at the
beginnig of the OpenXR session, if hand tracking support is detected
at runtime.