Initial UI for Web apps (#329)
This is the initial implementation of an UI for Web apps.

The user can add a new Web app by clicking on the "Save Web app…" button on the top bottom menu on the browser. This opens a dialog with the Web app's details. There is also a button in the URL bar with the same functionality.

Saved Web apps are displayed in a new tab in the Library.