Fix a crash in addons view
This is a regression caused by the upgrade to AC v75. The
SectionViewHolder component now requires a divider. We added the
code to retrieve the divider in Kotlin. However we didn't add
the divider in the resources file.

That's what we do now in order to provide a divider.

This fixes #341