Add the include path for OVR platform on Meta OpenXR builds
This is a leftover from b198a4c322a1291da7091660a4798acb0ce9c89a.
I forgot to add the CMake changes to that commit which add the
Meta VR platform includes to the Meta OpenXR build. That broke
the Meta OpenXR store builds.

We require them to perform the user entitlement.