Fix error pages regression
The upgrade to AC v75 caused a regression in error pages. They were
shown as meaningless sequences of characters to the user. The problem
was that AC removed the method we used to create the error pages, as
they're now using a different approach based on RequestInterceptor.

We selected createUrlEncodedErrorPage() as replacement by mistake,
that method does not generate an error page but an special URL that
is used by the RequestInterceptor.

In order to fix the issue we just proceed to read the HTML and CSS
resources we have for error pages and replace the placeholders with
the actual data from the request.

Fixes #336.