[OpenXR] Call xrBeginSession after resuming app only for HVR
The HVR platform has a very special way to handle the session. That
forced us to call xrRequestExitSession() when LeaveVR is called.
That happens for example when exiting the security zone. This meant
that we have to call xrBeginSession again when the application is
resumed. However that is not the case of the other OpenXR
implementations. For example the Oculus one does not follow that
pattern and thus it was crashing because we were trying to begin
a session that was not in READY state.

Fixes #358.