Add Pico XR build variant
This commit introduces a new build variant "picoxr" using Pico's OpenXR
SDK, and makes necessary changes to get Wolvic to run on the Pico 4. It
is likely that the variant also works with at least some versions of the
Pico Neo 3, but that was not tested.

This commit does not contain images or 3D models of the Pico 4
controllers, using empty placeholders instead. These might get replaced
in the future.

Furthermore, this commit is focussed on the Pico 4; later versions might
opt to differentiate between different Pico XR devices to permit better
adaptation to individual device models.