Fix WebXR controller profiles on Pico XR
This commit sets the controller type for Gecko for Pico XR controllers
to pico-neo2, causing them to get a suitable profile. It also adapts the
input mappings to match the profiles exposed through Gecko for consistency.

If Gecko should get updated to support more recently added types, we might
want to upgrade to pico-neo3 or, if available until then, a suitable
controller type for Pico 4.