Add option to adjust render resolution (OpenXR)
This commit introduces a new option to permit end-users to set a factor
scaling the render resolution used in OpenXR-based builds. Users that
primarily consume text content might want to increase this factor for
more clarity, while users that primarily run complex WebXR experiences
may want to lower it for better performance.

For now, this commit does not change the default values: Pico XR remains
at a factor of 1.4, while other builds default to 1.0. We might want to
discuss and define more balanced default values in a separate commit
later on.