[OpenXR] Vertically flip OpenXR layers in Meta
Most OpenXR layers in Meta appear upside down, i.e., flipped in
the y direction. In particular we're talking about Quad, Cylinder
and Equirect layers (cubemaps are not affected).

This does not happen in other devices using XRLayers (like Pico)
so it must be definitely a Meta implementation issue. In order
to fix that we use the XR_FB_composition_layer_image_layout extension
developed precisely by Meta.

We were using this only for equirect layers but it turns out that
quad and cylinder are also affected, so we've generalized the
same fix for the other layer types.

This allows us to remove a weird hack in Quad layers, that were
properly shown because the code was passing a negative value
for the layer height.