[OpenXR][Pico] Remove scaling factor (#387)
When the PicoXR flavour was added it came with some upscaling code
that increases the resolution used to render the Wolvic widgets.

However that's a mistake. We should not ignore the limits provided
by OpenXR as we might compromise stability. In fact this code
causes a lot of crashes (OOM issues) when using XRLayers as
the code will create swapchains bigger than expected.