[OpenXR][Pico] Enable hand tracking in the manifest
In order to launch Wolvic when the user is using hand tracking instead
of the controllers, we have to declare in the manifest that the
application supports hand tracking.

Note that the hand tracking support is still WIP, so don't expect it
to work on Pico just because we allow it in the manifest.