OpenXR: Draw hands by using a sphere for each hand joint
This is a stop-gap version of hand rendering that paints a ball
(shpere) for each hand joint detected.

While we aim at painting the hand mesh as provided by
XR_FB_hand_tracking_mesh, this is an initial, fallback version
that could be shown together, or as default on devices that don't
support the Facebook-specific extensions and only provide
XR_EXT_hand_tracking.