[Pico][XRLayers] Add a workaround for compressed textures
Pico OpenXR SDK does not support compressed textures format for layers.
In order not to get a crash we initialize the format to GL_RGBA8 which
is indeed supported by the runtime.

This means that when using layers, we are indeed still not able to
render the cubemap of the skybox (everything will be pitch black).