[Pico][XRLayers] Add a workaround for android surface layers
The specs[1] state that all members of the XrSwapchainCreateInfo structure
should be initialized to 0 before calling the method that creates the
swapchain. However in the case of Pico, doing that involves getting a
validation failure from the SDK.

In order to circumvent that we initialize the struct as if we don't
have layers support for the case of Pico.

[1] https://www.khronos.org/registry/OpenXR/specs/1.0/man/html/xrCreateSwapchainAndroidSurfaceKHR.html#XR_KHR_android_surface_swapchain