[XRLayers] Add support for cylinder layers
OpenXR cylinder layers were never properly supported despite having
code to handle them. There were several issues. First the layer was
sized 0x0 because the aspect ratio was not properly computed. Secondly
the radius of the cylinder was not properly set. And finally, the
pose was completely flawed, neither the position nor the orientation
were correct.

Regarding this last point, we have decided to add a new method to
pass the rotation to the cylinder layer because we were unable to
get it from the model matrix. I guess it should be possible but
this is the only way we found to make it work.