[Pico][XRLayers] Translate layers to average height
When using layers, we create a special space called layersSpace that
is translated -kAverageHeight in order to get the content centered on
user's face instead of the origin of coordinates.

However it looks like Pico's SDK does not properly use the layers space
when positioning the layer, forcing us to add a translation to individual
layers.